Fix: 식별자 형 바꿈 Integer > Long

+++ backend-spring/src/main/java/com/ssafy/pcgg/domain/recommend/entity/GpuEntity.java
@@ -5,14 +5,13 @@ import org.hibernate.annotations.CreationTimestamp;
 
 import java.math.BigDecimal;
 import java.time.LocalDate;
-import java.util.List;
 
 @Entity
 @Table(name = "part_gpu")
 public class GpuEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     @Column(length = 100)
     private String name;
